Combined pages 5 and 6 into one page.
