+++ client/index.jsx
@@ -26,7 +26,6 @@ import Page2 from './components/formPages/Page2';
 import Page3 from './components/formPages/Page3';
 import Page4 from './components/formPages/Page4';
 import Page5 from './components/formPages/Page5';
-import Page6 from './components/formPages/Page6';
 import ItineraryPage from './components/itineraryComponents/ItineraryPage';
 // import Register from './components/Register';
 
@@ -72,7 +71,6 @@ createRoot(root).render(
               <Route path="/form/page3" element={<Page3 />} />
               <Route path="/form/page4" element={<Page4 />} />
               <Route path="/form/page5" element={<Page5 />} />
-              <Route path="/form/page6" element={<Page6 />} />
             </Route>
 
             {/* Itinerary route */}
